refactor(): overall cleanups & refactoring (#31)
* feat(design): added utility to get more user friendly design names
* chore(actions): added back updateDesignFromParams action,
modified related code
* feat(ui): added basics of options display
* refactor(design): extracted design reducers & some of the export utils
* removed obsolete code
* moved all generic utilities under a 'utils' folder
* extracted io (export) utils, actions & reducers
* refactor(core): moved all of the 'core' code to 'design'
* split up scriptLoading.js into multiple specialized modules
* created a 'code-loading' folder & put relevant code there
* same for 'parameters'
* refactor(viewer): extracted viewer actions & reducers
* refactor(viewer): various tweaks
* regrouped all 'core' code that is ui independant
* regrouped all 'ui' code (includes actions & reducers)
* chore(styling): styling tweaks , added fonts
* chore(actions): made fetching the controls for parameter extraction more coherent
* fix(scriptLoading): added missing api function
* fix(prepareOutput): fixed bad import
* for now hiding options screen (as it is not functional yet)
