+++ packages/jscad-desktop/data/theme.light.json
@@ -8,5 +8,5 @@
       "color": [0.1, 0.1, 0.1, 0.5]
     }
   },
-  "mainTextColor": "#000"
+  "mainTextColor": "#424242"
 }
