+++ packages/jscad-desktop/src/core/code-loading/registerJscadExtension.js
@@ -1,0 +1,11 @@
+const fs = require('fs')
+const stripBom = require('strip-bom')
+
+const registerJscadExtension = () => {
+  require.extensions['.jscad'] = function (module, filename) {
+    const content = fs.readFileSync(filename, 'utf8')
+    module._compile(stripBom(content), filename)
+  }
+}
+
+module.exports = registerJscadExtension
