+++ packages/jscad-desktop/src/core/parameters/getParameterDefinitionsCLI.js
