+++ packages/jscad-desktop/src/sideEffects/titleBar.js
@@ -1,5 +1,4 @@
 function titleBarSink (outToTitle$) {
-
   outToTitle$.forEach(title => {
     document.title = title
   })
